some refactoring and inclusion of placeholder menu
