+++ solve.spec.py
@@ -1,5 +1,5 @@
 import unittest
-from main import *
+from solve import Sudoku
 
 test_sudoku = [[5,3,0,0,7,0,0,0,0],
           [6,0,0,1,9,5,0,0,0],
